+++ tests/testthat/test-readBindingConstraints.R
@@ -50,6 +50,7 @@ test_that("test if exist data value file", {
 
 # read latest version of empty study
 study_empty_latest_version <- setup_study_empty(sourcedir_empty_study)
+
 opts_test_empty <- antaresRead::setSimulationPath(study_empty_latest_version,
                                             "input")